objs bien passe

+++ cl/base.h
@@ -42,8 +42,8 @@ typedef struct		s_env
 typedef struct		s_obj
 {
 	int				type;
-	float3		pos;
-	float3		dir;
+	float3			pos;
+	float3			dir;
 	size_t			color;
 	double			alpha;
 	double			rayon;
@@ -67,8 +67,8 @@ typedef struct		s_scene
 
 struct		s_result_hit
 {
-	double			dist;
-	double			t;
+	float			dist;
+	float			t;
 	float3			norm; //contient le vecteur normal a la surface
 	float3			intersect; //contient le point dans le plan non translate d'intersection
 	t_obj			*obj; //pointeur sur lobjet intersecter
